requirements.txt
behave==1.2.6
Faker
allure-behave